chore: Party, PartyMember 엔티티에 ocid 추가 chore: PartyService 및 PartyMemberService에 파티장 예외사항 추가, feat: PartyController, PartyMemberController 구현

+++ src/main/java/project/maple/service/PartyService.java
@@ -4,6 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import project.maple.domain.Party;
+import project.maple.repository.PartyMemberRepository;
 import project.maple.repository.PartyRepository;
 
 import java.util.List;
@@ -14,6 +15,7 @@ import java.util.List;
 public class PartyService {
 
     private final PartyRepository partyRepository;
+    private final PartyMemberRepository partyMemberRepository;
 
     /*
     파티 생성
@@ -41,7 +43,29 @@ public class PartyService {
     파티 삭제
      */
     @Transactional
-    public void deleteParty(Long partyId) {
+    public void deleteParty(Long partyId, String userEmail) {
+        if (isPartyLeader(partyId, userEmail)) {
+            throw new IllegalArgumentException("파티장만 파티를 삭제할 수 있습니다.");
+        }
         partyRepository.deleteById(partyId);
+    }
+
+    /*
+    파티장 확인 메서드
+     */
+    public boolean isPartyLeader(Long partyId, String userEmail) {
+        Party findParty = partyRepository.findById(partyId).orElseThrow(() -> new IllegalArgumentException("존재하지 않는 파티입니다."));
+        return findParty.getMember().getUserEmail().equals(userEmail);
+    }
+
+    public void handleLeaderLeaving(Long partyId, String userEmail) {
+        // 1. 파티의 다른 멤버가 있는지 확인
+        boolean hasOtherMembers = partyMemberRepository.existsById(partyId);
+        if (hasOtherMembers) {
+            throw new IllegalStateException("파티장이 탈퇴할 수 없습니다. 다른 멤버가 존재합니다.");
+        }
+
+        // 2. 파티 삭제
+        deleteParty(partyId, userEmail);
     }
 }
